requirements.txt
Django>=4.2.0
django-ckeditor>=6.7.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
Pillow>=10.0.0
python-docx>=0.8.11
reportlab>=4.0.0
django-cors-headers>=4.2.0
openpyxl>=3.1.2
python-decouple>=3.8
django-redis>=5.3.0
redis>=4.6.0
django-filter>=23.5
weasyprint
reportlab
beautifulsoup4
xhtml2pdf
PyPDF2
bcrypt==3.2.0
pandas
openpyxl
django-widget-tweaks
django-crispy-forms
djangorestframework-simplejwt
